+++ src/assets/bars/src/scripts/types/Craft.d.ts
@@ -2,7 +2,6 @@
  * An instance of Craft.
  */
 declare const Craft: {
-  Commerce: any
   MatrixInput: GarnishComponent
   Smith: any
   t: (category: string, message: string, params?: object) => string